本地更改：修复方舟建设逻辑、夜晚行动移除建设方舟、玩家状态描述更新、天数联动、移除Google Fonts等

+++ src/main/java/com/example/snowisland/controller/ArkController.java
@@ -21,8 +21,8 @@ public class ArkController {
 
     @PostMapping("/invest")
     public Map<String, Object> investResources(
-            @RequestParam(required = false, defaultValue = "0") Integer wood,
-            @RequestParam(required = false, defaultValue = "0") Integer metal,
+            @RequestParam(required = false, defaultValue = "0") Double wood,
+            @RequestParam(required = false, defaultValue = "0") Double metal,
             @RequestParam(required = false, defaultValue = "0") Integer sealant
     ) {
         return arkService.investResources(wood, metal, sealant);
